Refactor modals to Modals folder and create Modals.css
