--- library-valcon-internship/src/components/ModalAddBook/ModalAddBook.css
@@ -1,1 +1,0 @@
-.ModalAddBook {}
