--- library-valcon-internship/src/components/ModalAddBook/ModalAddBook.tsx
@@ -1,5 +1,0 @@
-import './ModalAddBook.css'
-
-const ModalAddBook = () => <div className='ModalAddBook'>ModalAddBook Component</div>
-
-export default ModalAddBook
