+++ library-valcon-internship/src/components/Modals/ModalAddAuthor/ModalAddAuthor.css
@@ -1,41 +1,3 @@
-.modal {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  height: inherit;
-  width: 100%;
-  position: fixed;
-  top: 0;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  z-index: 140;
-}
-
-.overlay {
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.5);
-  position: fixed;
-  top: 0;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  z-index: 150;
-  backdrop-filter: blur(2px);
-}
-
-.content {
-  text-align: center;
-  background-color: white;
-  z-index: 160;
-  padding: 32px;
-  box-shadow: 0 1px 12px rgba(0, 0, 0, 0.5);
-  display: flex;
-  flex-direction: column;
-  border-radius: 8px;
-}
-
 .add-author-form {
   display: flex;
   flex-flow: column;
@@ -50,10 +12,6 @@
 
 .add-author-button {
   background-color: rgb(199, 149, 10);
-}
-
-.close-modal-btn {
-  background-color: red;
 }
 
 .modal-btn {
