+++ library-valcon-internship/src/components/Modals/Modals.css
@@ -1,0 +1,41 @@
+.modal {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    height: inherit;
+    width: 100%;
+    position: fixed;
+    top: 0;
+    bottom: 0;
+    left: 0;
+    right: 0;
+    z-index: 140;
+  }
+  
+  .overlay {
+    width: 100%;
+    height: 100%;
+    background-color: rgba(0, 0, 0, 0.5);
+    position: fixed;
+    top: 0;
+    bottom: 0;
+    left: 0;
+    right: 0;
+    z-index: 150;
+    backdrop-filter: blur(2px);
+  }
+  
+  .content {
+    text-align: center;
+    background-color: white;
+    z-index: 160;
+    padding: 32px;
+    box-shadow: 0 1px 12px rgba(0, 0, 0, 0.5);
+    display: flex;
+    flex-direction: column;
+    border-radius: 8px;
+  }
+
+  .close-modal-btn {
+    background-color: red;
+  }
