+++ library-valcon-internship/src/components/Navbar/Navbar.tsx
@@ -1,6 +1,9 @@
+import { useState } from 'react'
+
 import hamburgerMenuIcon from '../../assets/icons/hamburger-menu-icon.svg'
 import home from '../../assets/icons/home-icon.svg'
 import profile from '../../assets/icons/profile-icon.svg'
+import ModalAddBook from '../Modals/ModalAddBook/ModalAddBook'
 import NavbarLink from '../NavbarLink/NavbarLink'
 import './Navbar.css'
 
@@ -10,6 +13,7 @@ interface NavbarProps {
 }
 
 const Navbar = ({ handleMenuClick, token }: NavbarProps) => {
+  const [ showAddBookModal, setShowAddBookModal ] = useState(false)
   const isLoggedIn = token
   return isLoggedIn ? (
     <nav className='navbar'>
@@ -22,7 +26,7 @@ const Navbar = ({ handleMenuClick, token }: NavbarProps) => {
       <button className='navbar-link' onClick={() => handleMenuClick()}>
         <img src={hamburgerMenuIcon} alt='moreOptions-icon' />
       </button>
-      <button className='navbar-options-link'>
+      <button className='navbar-options-link' onClick={() => setShowAddBookModal(true)}>
         Add a new book
       </button>
       <NavbarLink to='option2' className='navbar-options-link' color='#F56211'>
@@ -31,6 +35,7 @@ const Navbar = ({ handleMenuClick, token }: NavbarProps) => {
       <NavbarLink to='option3' className='navbar-options-link' color='#F56211'>
         Option 3
       </NavbarLink>
+      <ModalAddBook show={showAddBookModal} closeModal={() => setShowAddBookModal(false)} />
     </nav>
   ) : (
     <></>
